Move package `molecule` into package `type`
