+++ types/molecule/type.go
@@ -34,7 +34,6 @@ func (b *Byte) AsSlice() []byte {
 	return b[:]
 }
 
-//
 func ByteFromSlice(slice []byte, _compatible bool) (*Byte, error) {
 	if len(slice) != 1 {
 		return nil, errors.New("TotalSizeNotMatch")
